BLOC listener instead of FutureBuilder is SplashScreen
