+++ lib/screens/auth/signup/signup_screen.dart
@@ -1,8 +1,8 @@
-import 'package:authentication_repository/authentication_repository.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gp5/extensions/build_context_extensions.dart';
 import '../../../enums/status_enum.dart';
+import '../../../repos/authentication/authentication_repository.dart';
 import '../../../routes/app_routes.dart';
 import 'bloc/signup_bloc.dart';
