+++ lib/screens/home/home_screen.dart
@@ -1,8 +1,8 @@
-import 'package:authentication_repository/authentication_repository.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gp5/routes/app_routes.dart';
 
+import '../../repos/authentication/authentication_repository.dart';
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
